requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.4
python-multipart>=0.0.9
httpx>=0.27.0
pandas>=2.2.0
sqlalchemy>=2.0.35
aiosqlite>=0.20.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
apscheduler>=3.10.4
beautifulsoup4>=4.12.3
lxml>=5.3.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
